requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.15.1
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.9.0
asyncer==0.0.8
attrs==24.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
dataclasses-json==0.6.7
datasets==3.4.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.6.13
dspy-ai==2.6.13
dydantic==0.0.7
fastapi==0.115.12
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
greenlet==3.1.1
grpcio==1.62.1
grpcio-tools==1.62.1
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.40.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-community==0.3.13
langchain-core==0.3.47
langchain-mcp-adapters==0.0.5
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.2.70
langgraph-checkpoint==2.0.10
langgraph-checkpoint-postgres==2.0.13
langgraph-checkpoint-sqlite==2.0.1
langgraph-sdk==0.1.51
langsmith==0.2.11
litellm==1.60.0
loguru==0.7.3
magicattr==0.1.6
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.25.1
mcp==1.4.1
ml-dtypes==0.4.1
mmh3==4.1.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.59.8
optuna==4.2.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
ply==3.11
propcache==0.2.1
protobuf==4.25.7
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2025.1
PyYAML==6.0.2
redis==5.0.1
redisvl==0.5.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.2
setuptools==80.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.23
sse-starlette==2.2.1
starlette==0.41.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
trustcall==0.0.26
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
wget==3.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
pandas
pyarrow
pinecone-client[grpc]
google-api-python-client
aioboto3
langchain-groq
